Begin making study happy without specimen module around: register SpecimenRequestNotificationEmailTemplate from specimen module, move CompleteType out of SpecimenService, check SpecimenService.get() != null
Move AbstractStudyPipelineJob to study-api
