+++ api/src/org/labkey/api/study/CompletionType.java
@@ -1,0 +1,9 @@
+package org.labkey.api.study;
+
+public enum CompletionType
+{
+    SpecimenGlobalUniqueId,
+    ParticipantId,
+    VisitId,
+    LabId
+}
