+++ api/src/org/labkey/api/study/query/PublishResultsQueryView.java
@@ -51,6 +51,7 @@ import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QuerySettings;
 import org.labkey.api.reports.ReportService;
 import org.labkey.api.security.User;
+import org.labkey.api.study.CompletionType;
 import org.labkey.api.study.ParticipantVisit;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.Study;
@@ -821,7 +822,7 @@ public class PublishResultsQueryView extends ResultsQueryView
         protected ActionURL getCompletionBase(RenderContext ctx)
         {
             Container c = rowTargetStudy(_resolverHelper, ctx);
-            return SpecimenService.get().getCompletionURL(c, SpecimenService.CompletionType.ParticipantId);
+            return SpecimenService.get().getCompletionURL(c, CompletionType.ParticipantId);
         }
 
         @Override
@@ -843,7 +844,7 @@ public class PublishResultsQueryView extends ResultsQueryView
         protected ActionURL getCompletionBase(RenderContext ctx)
         {
             Container c = rowTargetStudy(_resolverHelper, ctx);
-            return SpecimenService.get().getCompletionURL(c, SpecimenService.CompletionType.VisitId);
+            return SpecimenService.get().getCompletionURL(c, CompletionType.VisitId);
         }
 
         @Override
