+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -36,10 +36,12 @@ import org.labkey.api.specimen.model.SpecimenDomainKind;
 import org.labkey.api.specimen.model.SpecimenEventDomainKind;
 import org.labkey.api.specimen.model.SpecimenRequestEventType;
 import org.labkey.api.specimen.model.VialDomainKind;
+import org.labkey.api.specimen.view.SpecimenRequestNotificationEmailTemplate;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.importer.SimpleStudyImporterRegistry;
 import org.labkey.api.study.writer.SimpleStudyWriterRegistry;
+import org.labkey.api.util.emailTemplate.EmailTemplateService;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.WebPartFactory;
 import org.labkey.specimen.action.SpecimenApiController;
@@ -118,6 +120,7 @@ public class SpecimenModule extends SpringModule
         // Register early -- some modules don't declare a runtime dependency on specimen module, but will use the
         // service if it's available
         SpecimenService.setInstance(new SpecimenServiceImpl());
+        EmailTemplateService.get().registerTemplate(SpecimenRequestNotificationEmailTemplate.class);
     }
 
     @Override
