+++ study/src/org/labkey/study/model/DatasetDefinition.java
@@ -101,6 +101,7 @@ import org.labkey.api.security.permissions.UpdatePermission;
 import org.labkey.api.security.roles.RoleManager;
 import org.labkey.api.security.roles.SiteAdminRole;
 import org.labkey.api.settings.AppProps;
+import org.labkey.api.study.CompletionType;
 import org.labkey.api.study.Dataset;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.Study;
@@ -1213,7 +1214,7 @@ public class DatasetDefinition extends AbstractStudyEntity<DatasetDefinition> im
     {
         private final ActionURL _completionBase;
 
-        public AutoCompleteDisplayColumnFactory(Container studyContainer, SpecimenService.CompletionType type)
+        public AutoCompleteDisplayColumnFactory(Container studyContainer, CompletionType type)
         {
             _completionBase = SpecimenService.get().getCompletionURL(studyContainer, type);
         }
@@ -1343,7 +1344,7 @@ public class DatasetDefinition extends AbstractStudyEntity<DatasetDefinition> im
 
             var sequenceNumCol = newDatasetColumnInfo(this, getStorageColumn("SequenceNum"), getSequenceNumURI());
             sequenceNumCol.setName("SequenceNum");
-            sequenceNumCol.setDisplayColumnFactory(new AutoCompleteDisplayColumnFactory(_container, SpecimenService.CompletionType.VisitId));
+            sequenceNumCol.setDisplayColumnFactory(new AutoCompleteDisplayColumnFactory(_container, CompletionType.VisitId));
             sequenceNumCol.setMeasure(false);
 
             if (def.isDemographicData())
